+++ src/icemac/ab/calendar/browser/tests/test_startpage.py
@@ -7,7 +7,7 @@ class StartpageDispatchSTests(icemac.ab.calendar.testing.SeleniumTestCase):
     def test_startpage_redirects_to_calendar_if_set_on_address_book(self):
         self.login()
         sel = self.selenium
-        sel.open('/ab/@@edit.html')
+        sel.open('/ab/@@edit-address_book.html')
         sel.select('id=form-widgets-startpage', 'label=Calendar')
         sel.type('id=form-widgets-title', 'Test')
         sel.clickAndWait('id=form-buttons-apply')