+++ CHANGES.rst
@@ -2,11 +2,13 @@
  Changelog
 ===========
 
-1.4.1 (unreleased)
+1.5.0 (unreleased)
 ==================
 
 - The calendar tab in the main menu is now highlighted if the calendar or an
   event is displayed.
+
+- Adapt tests and code to changes in `icemac.addressbook 2.5`.
 
 
 1.4.0 (2014-03-07)